github-pages-deploy-action-temp-deployment-folder/requirements.txt
numpy
pytest-html
allure-pytest
